Create _masthead and _drawerbanner partials, plus sections/index
This commit sees the creation of the masthead and drawerbanner
partials for the main application.html.erb template, plus a
first attempt at filling in article boxes on the sections/index
view to see if we can get anything to show. Expecting many
changes to follow these initial set-ups before we have something
working.