Add login/logout top banner, plus date from live data
We made the decision to include the login/logout options at the
very top of the page instead of the drawer, so that we could
include an ad for why you should create an account if you never
have. Wrapped it all in a dark orange banner to get it to
stand out more. Also replaced the wireframe date in the
masthead with a live Time.now pull of real data.