Create new "page_name" variable for system-wide use
Big new decision -- a @page_name variable for defining the text
that will appear in the yellow bar at the top of every page.
For section views, @page_name will be set to @section.name; for
pages like "Create Account" and "Edit Subscription," it'll be
hard-coded right in the controller.