Add login form to top orange banner on every page
Instead of a separate login page, we decided to include it as a
mini-form in the top banner on every page, showing up whenever
a person is not currently logged in. This also runs with a little
ad for why you should create an account if you never have, and
a link to go do so (which DOES lead to a separate page).