Move drawer code into sub-partial
Another major new decision -- we will be rendering a separate
template for those who are not logged in versus those who are,
and the non-logged-in template should not show the drawer for
choosing specific sections, editing subscriptions, etc. Since
the drawer code had been married to the banner code in the
'drawerbanner' partial, it had to be pulled out into its own
sub-partial, and a new partial made called "drawernobanner" for
the non-logged-in version. Part of the previous application.html.erb
template now has to be moved into individual pages as well, so
to call either for the partial with the drawer or without; this
also involves repeating two lines of setting up the "content
dividers" for rendering the entire blue part of the page, so
these two lines were moved into their own "main" partial as
well, for DRY sake.