Establish initial sections/default view coding plus partials
After some trial and error, I decided the best way to render
articles correctly on a page, based on how popular they are,
is to assign a variable value of "small," "medium," "large" or "splash"
to them, then call a case statement on that variable in our
view template and render one of several partials based on
its value. This update sees the initial creation of those
partials, plus an initial case statement on our sections/default
view (the general one that will show if a person isn't logged
in); but there's a LOT of work to do with Emmanuel this afternoon
as far as getting all the references to backend variables correct,
as well as this new "article.column_class" variable added.