New names to "show stories" partials, plus many small cleanups
After a group walkthrough of our "show stories" templates, we've
decided on a naming convention for the variables being referenced;
these partials and and the "sections/default" template now reflect
them. I also added a @page_name declaration in the "default"
section controller, and tweaked the styling of the revamped
"create user" form.